backend/requirements.txt
django>=5.2,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
dj-database-url>=2.1,<3.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
gunicorn>=22.0,<24.0
whitenoise>=6.6,<7.0
django-cors-headers>=4.3,<5.0
resend>=2.26,<3.0
